feature: refactor offline MP4 runtime into reusable inference service with shared runtime primitives
